게시물 수정, 삭제 시 로그인 체크

+++ src/main/java/com/example/sbs/cuni/controller/ArticleController.java
@@ -3,6 +3,8 @@ package com.example.sbs.cuni.controller;
 import java.util.List;
 import java.util.Map;
 
+import javax.servlet.http.HttpSession;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
@@ -39,7 +41,20 @@ public class ArticleController {
 	}
 
 	@RequestMapping("article/modify")
-	public String showModify(Model model, int id) {
+	public String showModify(Model model, HttpSession session, int id) {
+		int loginedMemberId = 0;
+
+		if (session.getAttribute("loginedMemberId") != null) {
+			loginedMemberId = (int) session.getAttribute("loginedMemberId");
+		}
+
+		if (loginedMemberId == 0) {
+			model.addAttribute("alertMsg", "로그인 후 이용해주세요.");
+			model.addAttribute("historyBack", true);
+
+			return "common/redirect";
+		}
+
 		Article article = articleService.getArticle(id);
 
 		model.addAttribute("article", article);
@@ -48,7 +63,21 @@ public class ArticleController {
 	}
 
 	@RequestMapping("article/doModify")
-	public String doModify(Model model, @RequestParam Map<String, Object> param) {
+	public String doModify(Model model, @RequestParam Map<String, Object> param, HttpSession session) {
+
+		int loginedMemberId = 0;
+
+		if (session.getAttribute("loginedMemberId") != null) {
+			loginedMemberId = (int) session.getAttribute("loginedMemberId");
+		}
+
+		if (loginedMemberId == 0) {
+			model.addAttribute("alertMsg", "로그인 후 이용해주세요.");
+			model.addAttribute("historyBack", true);
+
+			return "common/redirect";
+		}
+
 		Map<String, Object> rs = articleService.modify(param);
 
 		int id = Integer.parseInt((String) param.get("id"));
@@ -63,7 +92,21 @@ public class ArticleController {
 	}
 
 	@RequestMapping("article/doDelete")
-	public String doDelete(Model model, int id) {
+	public String doDelete(Model model, int id, HttpSession session) {
+
+		int loginedMemberId = 0;
+
+		if (session.getAttribute("loginedMemberId") != null) {
+			loginedMemberId = (int) session.getAttribute("loginedMemberId");
+		}
+
+		if (loginedMemberId == 0) {
+			model.addAttribute("alertMsg", "로그인 후 이용해주세요.");
+			model.addAttribute("historyBack", true);
+
+			return "common/redirect";
+		}
+
 		Map<String, Object> rs = articleService.deleteArticle(id);
 
 		String msg = (String) rs.get("msg");
